+++ src/Infrastructure/Console/Commands/ShowFieldsDescriptionCommand.php
@@ -22,11 +22,13 @@ use Bitrix24\SDK\Core\Response\Response;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Helper\QuestionHelper;
 use Symfony\Component\Console\Helper\Table;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Question\ChoiceQuestion;
+use Symfony\Component\Console\Question\Question;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
 
@@ -102,7 +104,9 @@ class ShowFieldsDescriptionCommand extends Command
                     $fieldsMethods[] = $method;
                 }
             }
-
+            /**
+             * @var QuestionHelper $helper
+             */
             $helper = $this->getHelper('question');
             $itemQuestion = new ChoiceQuestion(
                 'Please select item number to see fields',